feat: Add functions to retrieve object icons
This commit moves the symbol factories (everything ending with "Symbol")
from the SVG module's `deco_factories` to their own submodule below
`capellambse.diagram`. New functions are added that expose the symbols
as standalone SVG images with configurable size.
