+++ capellambse/diagram/__init__.py
@@ -10,6 +10,7 @@ from ._vector2d import *
 
 from .capstyle import *
 from ._diagram import *
+from ._icons import *
 from ._json_enc import *
 from ._styleclass import *
 
@@ -19,10 +20,11 @@ if not t.TYPE_CHECKING:
     from ._vector2d import __all__ as _all1
     from .capstyle import __all__ as _all2
     from ._diagram import __all__ as _all3
-    from ._json_enc import __all__ as _all4
-    from ._styleclass import __all__ as _all5
+    from ._icons import __all__ as _all4
+    from ._json_enc import __all__ as _all5
+    from ._styleclass import __all__ as _all6
 
-    __all__ = [*_all1, *_all2, *_all3, *_all4, *_all5]
+    __all__ = [*_all1, *_all2, *_all3, *_all4, *_all5, *_all6]
 
-    del _all1, _all2, _all3, _all4, _all5
+    del _all1, _all2, _all3, _all4, _all5, _all6
 del t
